Refactored newsletter form into separate component
